Refactor as Listener
